backend/requirements.txt
# Flask and related packages
Flask==2.2.3
Flask-Cors==3.0.10
Werkzeug==2.2.3

# Database
pymongo==4.3.3

# Machine Learning packages
scikit-learn==1.2.2
pandas==1.5.3
numpy==1.24.3
joblib==1.2.0

# Google Generative AI
google-generativeai==0.3.2

# Data validation
pydantic==2.5.2

# JSON handling
typing-extensions==4.9.0

# Attendence
opencv-python==4.5.5.64
face-recognition==1.3.0
dlib==19.24.6
Pillow==9.3.0